refactor: extract applyWorldStateThreadUpdates from narrativeStore
narrativeStore.ts was 1,465 lines. Move applyWorldStateThreadUpdates
(handles per-segment updates to player-character threads, character
relationships, and major-event checkpoints on the world store) plus its
WorldStateUpdateParams interface, lazy characterStoreModule cache, and
SEGMENT_SNIPPET_MAX_LENGTH constant into a dedicated
narrativeStore.worldStateThreads.ts module — same naming pattern as
loreStore.actions.*.ts / loreStore.helpers.ts.

Replace the previous empty `catch {}` at the function boundary with a
logger.warn so background failures are at least observable; behavior is
otherwise byte-for-byte identical.

narrativeStore.ts: 1,465 -> 1,241 lines. First slice of a broader split;
the store still owns segments/decisions/endings/persistence and needs
further breakup.
